Add files via upload
The purpose of this assignment was to create a virtual experiment, designed to test two virtual machines. This allowed us in a (realistic) way to learn how to work as a Network Administrator. We had to install features that are in Windows to get our devices and network up and running.
Working and doing tests helped to develop the knowledge that security is very important. For example, one tool, the firewall, is essential to prevent malicious access to any network. Also, when you are going to give access to the user, you shouldn’t give too much permission to them in a company or college.